requirements.txt
# Django
# ------------------------------------------------------------------------------
django==4.2.5  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.56.1  # https://github.com/pennersr/django-allauth
# django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
# crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
# django-redis==5.3.0  # https://github.com/jazzband/django-redis

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.2.0  # https://github.com/adamchainz/django-cors-headers

# DRF-spectacular for api documentation
drf-spectacular==0.26.4  # https://github.com/tfranzel/drf-spectacular


# ADD
# ------------------------------------------------------------------------------
djangorestframework-simplejwt
django-filter==23.2

django-debug-toolbar==4.2.0  # https://github.com/jazzband/django-debug-toolbar